requirements.txt
aiohttp==3.7.4.post0
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cycler==0.10.0
discord.py==1.7.3
Django==3.2.2
et-xmlfile==1.1.0
ffmpeg-python==0.2.0
future==0.18.2
idna==2.10
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.4.2
multidict==5.1.0
numpy==1.20.2
openpyxl==3.0.7
pandas==1.2.4
Pillow==8.2.0
prettytable==2.0.0
pycparser==2.20
pymongo==3.11.4
PyNaCl==1.4.0
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt==5.15.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
riotwatcher==3.1.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0
typing-extensions==3.10.0.0
urllib3==1.26.4
wcwidth==0.2.5
xlwt==1.3.0
yarl==1.6.3
youtube-dl==2021.6.6
